file-transfers: add (admin) HTTP endpoints for modifying configs
Issue: https://github.com/moov-io/paygate/issues/147
