+++ go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/mattn/go-sqlite3 v1.10.0
 	github.com/moov-io/accounts v0.3.1-0.20190522191929-18df7fa8fa58
 	github.com/moov-io/ach v0.6.0-rc7.0.20190424165414-c1497c09b16f
-	github.com/moov-io/base v0.10.0-rc1
+	github.com/moov-io/base v0.10.0-rc1.0.20190612164034-af9fee782e7e
 	github.com/moov-io/fed v0.1.4
 	github.com/moov-io/ofac v0.7.1-0.20190506160459-5cf50ac71326
 	github.com/ory/dockertest v3.3.4+incompatible
